Refactoring, Created folders for game and main_menu and created a relevant mod file and plugins.
