+++ src/game/grid_traits.rs
