+++ src/lib.rs
@@ -1,15 +1,2 @@
-pub mod attack;
-pub mod building_spawning;
-pub mod buildings;
-pub mod camera;
-pub mod castle_fight_ldtk;
-pub mod grid_traits;
-pub mod health;
 pub mod inspector_plugin;
-pub mod movement;
-pub mod resources;
-pub mod teams;
-pub mod unit_spawning;
-pub mod units;
-pub mod vision;
-pub mod waypoints;
+pub mod game;
