+++ src/main_menu/mod.rs
@@ -1,0 +1,13 @@
+use bevy::prelude::*;
+
+pub struct MainMenuPlugin;
+
+impl Plugin for MainMenuPlugin {
+    fn build(&self, app: &mut App) {
+        app.add_systems(Startup, main_menu);
+    }
+}
+
+pub fn main_menu() {
+    println!("You are on the main menu");
+}
